+++ src/utilities/app-connection.js
@@ -6,6 +6,8 @@
  * Date: 02-13-2024
  *
  * Notes:
+ * ID02062025: ganrad: v2.2.0: Added new methods to a) Check if a connection exists for an Ai App & b) Delete an 
+ * Ai App Connection.
  *
 */
 class AppConnections {
@@ -33,5 +35,15 @@ class AppConnections {
   getAllConnections() {
     return this.appData;
   }
+
+  // ID02062025.sn
+  hasConnection(id) {
+    return this.appData.has(id);
+  }
+
+  removeConnection(id) {
+    return this.appData.delete(id); // Returns true if element was present
+  }
+  // ID02062025.en
 }
 module.exports = AppConnections;